Incorporate process_files into a processing script.
Reset java_ast to make it compatible. Later will need to refactor again.
